requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
fake-useragent==2.0.3
frozenlist==1.5.0
h11==0.14.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
msgspec==0.19.0
multidict==6.1.0
outcome==1.3.0.post0
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3